requirements.txt
pynput>=1.7.6
pyperclip>=1.8.2
pywin32>=306
pystray>=0.19.5
customtkinter>=5.2.0
Pillow>=10.0.0
pyinstaller>=6.0.0
pytest>=7.4.0
